refact: move authentication Koin module to :feature-authentication
