+++ app/src/shared/java/com/jeanbarrossilva/ongoing/app/OngoingApplication.kt
@@ -3,8 +3,8 @@ package com.jeanbarrossilva.ongoing.app
 import android.app.Application
 import com.jeanbarrossilva.ongoing.app.module.boundaryModule
 import com.jeanbarrossilva.ongoing.app.module.coreModule
-import com.jeanbarrossilva.ongoing.app.module.feature.authenticationModule
 import com.jeanbarrossilva.ongoing.app.module.feature.extensionsModule
+import com.jeanbarrossilva.ongoing.feature.authentication.authenticationModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.startKoin
 import org.koin.core.context.stopKoin
